% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Act2Daily.R
\name{Act2Daily}
\alias{Act2Daily}
\title{Extract Actigraphy Daily Recording}
\usage{
Act2Daily(
  df,
  Bdf,
  ID,
  TUnit = "hour",
  VAct = NULL,
  VTm = NULL,
  Incomplete = FALSE
)
}
\arguments{
\item{df}{A data.frame containing the actigraphy records}

\item{Bdf}{A BriefSum object (default NULL, i.e., not included). This is essential when time adjustment is needed.}

\item{ID}{The subject's ID which would be used to create a folder.}

\item{TUnit}{The unit of time to be displayed in the daily actigraphic recording exported. These include: "day" "hour" "minute" "second"  (default: "hour")}

\item{VAct}{The name of the variable containing activity score  (defult assumed the second variable)}

\item{VTm}{The name of the variable containing timestamp (default assumed the first variable)}

\item{Incomplete}{A binary operator to decide whether to keep any less than 24hr-long recordings. (default: FALSE)}
}
\description{
Extract Actigraphy Daily Recording
}
\keyword{Actigraphy}
\keyword{Daily}
\keyword{Separate}
