% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BriefSum.R
\name{BriefSum}
\alias{BriefSum}
\title{Summarize the Actigraphy Recording by Day}
\usage{
BriefSum(df, SR, Start, TZ = "local")
}
\arguments{
\item{df}{A data.frame containing the actigraphy records}

\item{SR}{The sampling rate of the actigraphy (unit at Hz). Note that Hz should be equal to or less than 1.}

\item{Start}{The starting date and time of the recording in formats like this "2021-03-05 18:31:03"}

\item{TZ}{The time zone when the recording started. (default = "local", which means user's local time zone)}
}
\description{
Summarize the Actigraphy Recording by Day
}
\keyword{Actigraphy}
\keyword{Summary,}
