% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DST2GL.R
\name{DST2GL}
\alias{DST2GL}
\title{Compute Time Gain/Loss due to Daylight Saving Time}
\usage{
DST2GL(Date, TZ = "local", DT = NULL)
}
\arguments{
\item{Date}{A vector of class \code{Date}, \code{POSIXct}, or \code{POSIXlt},
sorted in ascending order.  Each element marks the start of a day
boundary.  To capture DST transitions, include at least one day
before and after the expected shift. Note that the time zone should
be specified in the DT. See \code{as.POSIXct}.}

\item{TZ}{The time zone when the recording started. (default = "local",
which means user's local time zone)}

\item{DT}{`r lifecycle::badge("deprecated")` Use `Date` instead.}
}
\value{
A numeric vector of the same length as \code{Date}. Each
  value is the number of hours that day is longer (positive)
  or shorter (negative) than 24h. A zero means no DST shift
  on that date.
}
\description{
DST2GL computes, for each date or date-time entry, how many hours a day
deviates from the standard 24hours because of DST shifts.
Internally it calls \code{\link{Date2TotalT}} in seconds, subtracts
the 86400s of a normal day, and converts the remainder to hours. Note,
while **DST** is part of the function name, DST2GL intends to capture
also non-standard time shift.
}
\details{
Non-uniform Shift in Offset Values:
While around most daylight saving time (DST) transitions have
offset by an hour since 1980, various exceptions exists (e.g.,)


Output Signs and Coordinated Universal Time (UTC) Offsets:
During transitioning from standard time to DST, the clock
springs forward by a set number of minutes. This results in
a shorter day (e.g., 23h). In this case, the output of DST2GL will
be negative (e.g., -1) and vice versa.
}
\examples{
# Example of standard time to daylight saving time transition
# On 2021-03-14 clocks jumped forward: day is 23h so output = -1
DST2GL(Date = "2021-03-14", TZ = "America/New_York")

# Example around a fall-back transition (e.g. first Sunday in November)
dates <- as.Date("2021-11-07")

# On 2021-11-07 clocks fall back: day is 25h so output = +1
DST2GL(dates)

# Non-standard time shift
DST2GL(Date= "3/29/1992", TZ = "America/Guyana")


# Multiple Dates
sapply(c("2021-03-13", "2021-03-14", "2021-03-15"),
       DST2GL)

}
\seealso{
\code{\link{Date2TotalT}} \code{\link{DST}} \code{\link{TZwDST}}
}
\keyword{dst}
