% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FlyEast_adj}
\alias{FlyEast_adj}
\title{Adjusted Travel East Sample Data}
\format{
An adjusted FlyEast data with 40 thousands plus rows and
  3 variables:
\describe{
\item{DateTime}{The time coordinates of each recorded activies in
  the standard date and time joint format}
\item{Date}{The date of each recorded activies}
\item{Hour}{The simplified time coordinates in hour form converted
  from `DateTime`}
\item{Activity}{The activity count measured by a tri-axis
  actigraphy}
\item{Activity_ID}{Numeric values denoted the sequence of each
  activity count in the recording}
\item{Note}{Summary annotation derived from
  \code{\link{BriefSum}}.}
}
}
\usage{
data(FlyEast_adj)
}
\description{
This is the subset of the travel-adjusted `FlyEast` data. The data
was originally collected to monitor the influence of jetleg on
professional athelets' performance during competitions.
}
\seealso{
\code{\link{BriefSum}} \code{\link{TAdjust}} \code{\link{FlyEast}}
}
\keyword{FlyEast_adj}
\keyword{adjust}
