% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IANA}
\alias{IANA}
\title{IANA Time Table}
\format{
A modified IANA 2025b database
 \describe{
 \item{Country_Name}{Full name of the country (e.g., "Canada").}
 \item{Country_Code}{Two-letter ISO 3166-1 alpha-2 country code (e.g., "CA").
 }
 \item{Timezone_IANA}{Canonical IANA timezone identifier (e.g.,
 "America/Montreal").}
 \item{TimeZone_Identifiers}{Alternative or historical identifiers linked
 to the IANA timezone.}
 \item{Standard_Time_Zone_long_name}{Human-readable name of the timezone
 during standard time, without daylight saving adjustment (e.g., "Eastern
 Standard Time").}
 \item{Standard_TZ_Code}{Abbreviation used during standard
 (non-daylight saving) time only (e.g., "EST").}
 \item{Standard_Offset}{UTC offset in clock time during standard time,
 excluding daylight saving adjustment (e.g., +03:00).}
 \item{Standard_Offset_Numeric}{Numeric UTC offset in hours during standard
 time (e.g., 3.00).}
 \item{Observes_DST}{Logical flag indicating whether the timezone observes}
 \item{DST_Time_Zone_long_name}{Human-readable name of the timezone
 during daylight saving time or other recurrent practice of time shift
 (e.g., "Australian Central Daylight Time").}
 \item{DST_TZ_Code}{Abbreviation used during daylight saving time or other
 recurrent practice of time shift only (e.g., "ACDT").}
 \item{DST_Offset}{UTC offset in clock time during daylight saving or
 recurrent time change (e.g., +10:30).}
 \item{DST_Offset_Numeric}{Numeric UTC offset in hours during daylight
 saving or recurrent time change (e.g., 10.5)}
 }
}
\usage{
data(IANA)
}
\description{
A dictionary for standard abbreviation and labels of time-zone and their
corresponding UTC offset.
}
\seealso{
\code{\link{TZwDST}} \code{\link{MapTZ}}
}
\keyword{IANA}
