% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTZ.R
\name{MapTZ}
\alias{MapTZ}
\title{Build DST grouping table and leaflet map}
\usage{
MapTZ(Year = "Current", Table = FALSE)
}
\arguments{
\item{Year}{Integer year for DST calculation. Default is "Current", which
uses the current year of the system.}

\item{Table}{Logical; whether to return the DST grouping table
(default = FALSE).}
}
\value{
list with:
  - table: grouped TZ table
  - map: leaflet map (if sfTZ supplied)
}
\description{
This function builds a table grouping IANA time zones by their UTC offset
and DST transition dates/times for a specified year. If provided with a
\code{sf} object of time zone polygons, it also generates a leaflet map
coloring zones by their DST characteristics.
}
\examples{
# Build DST table and map for all time zones in 2024
if(interactive()) {
   MapTZ(Year = 2024) # Just the map without table
}

}
\seealso{
\code{\link{Geo2TZ}} \code{\link{TZwDST}} \code{\link{TZ2UTC}}
}
\keyword{country}
\keyword{map}
\keyword{plot}
\keyword{tz}
