% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mIANA.R
\name{mIANA}
\alias{mIANA}
\title{Mini IANA Table for Internal Uses}
\usage{
mIANA(Write = FALSE, Dir = NULL)
}
\arguments{
\item{Write}{A binary code to indicate whether to write a .csv file containing the template needed for the travel log. (default = FLASE)
When set to FALSE, the template will be returned as an object.
When set to TRUE, user must provide the target directory where they wish to store the template.}

\item{Dir}{The directory where the travel log template to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
}
\value{
a miniturized IANA table template as a data.frame or written as a CSV file
}
\description{
A miniturized IANA table generator that only extract the first timezone based
on the abbreviated time zone code. An internal function to resolve the internal
data access issue.
}
\examples{
\dontrun{

sIANA <- mIANA(Wirte = FALSE)

print(sIANA)

}

}
\keyword{Log}
\keyword{Shift}
\keyword{Template}
\keyword{Timezone}
\keyword{Travel}
