% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonWear.R
\name{NonWear}
\alias{NonWear}
\title{Detection for Possible Non-wearer Periods}
\usage{
NonWear(
  data,
  VAct = NULL,
  VTm = NULL,
  method = "Choi",
  HyperP = list(ScreenEpoch = 90, ArtEpoch = 2, TraceEpoch = 30)
)
}
\arguments{
\item{data}{A data.frame of raw actigraphy recording. Both time and activity
count should be included in the \code{data}. See \code{VAct} and \code{VTm}
for further detail.}

\item{VAct}{Optional character. Name of the activity column in \code{data}.
If NULL, defaults to the second column of \code{data}.}

\item{VTm}{Optional character. Name of the date.time index column in
\code{data}.If NULL, defaults to the first column of \code{data}.}

\item{method}{Character string specifying detection method
\itemize{
  \item "Choi": A modified Troiano algorithm designed to detect possible
  non-wear time using moving window approach. The algorithm considers the
  following hyperparameters:
   \itemize{
     \item ScreenEpoch: minimum threshold to classifying a segment as
     probable non-wear time; Default = 90
     (minutes per epoch).
     \item Artifact Control:
      \itemize{
       \item ArtEpoch: The maximum interval of spontaneous activity
       (artifact) allowed between two possible non-wear time periods.
       Default = 2 (minutes)
       \item TraceEpoch: The minimum time of inactivity for the upstream and
       downstream of the identified spontaneous activity to be considered
       the entire segment as possible non-wear. Default = 30 (minutes).
       }
       }
}}

\item{HyperP}{a list of hyperparameters used to tune the non-wear detection
algorithm. The default values are set for the Choi algorithm, which is
currently the only supported method.}
}
\value{
An amended data.frame with non-wearer detection results storing in a new
column:
  \item \code{pNonWear}, which stands for possible non-wear time. The values
  are logical, where \code{TRUE} indicates a possible non-wear period
  according to the specified method (e.g., Choi algorithm).
}
\description{
`NonWear()` is a function that detects possible non-wearing periods within
a longitudinal recording. Currently, it supports only the Choi algorithm and
require \eqn{1/60}Hz sampling rate.
}
\examples{

# Create quick summary of the recording with adjustment for daylight saving.
BdfList <-
  BriefSum (
  data = FlyEast,
  SR = 1 / 60,
  Start = "2017-10-24 13:45:00"
  )

data <- BdfList$data

# Detect possible non-wear time using the Choi algorithm with default
# hyperparameters (i.e., ScreenEpoch = 90, ArtEpoch = 2, TraceEpoch = 30)
data <- NonWear (data = data,
                 VAct = "Activity",
                 VTm = "Time",
                 method = "Choi")


# Visualize the detected non-wear time (red) against raw data (black)
## First create a new variable
plot (y = data$Activity, x = data$DateTime, type = "l")
points (y = data$pNonWear * 1000,
        x = data$DateTime,
        col = "red", pch = 16, cex = 0.5)

}
\references{
Choi L, Liu Z, Matthews CE, Buchowski MS. Validation of accelerometer wear
and nonwear time classification algorithm. Med Sci Sports Exerc. Feb 2011;
43(2):357-64. doi:10.1249/MSS.0b013e3181ed61a3
}
\seealso{
\code{\link[PhysicalActivity]{wearingMarking}}
}
\keyword{cosinor}
