% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num2UTC.R
\name{Num2UTC}
\alias{Num2UTC}
\title{Convert Numbers to Standard UTC Offset}
\usage{
Num2UTC(x)
}
\arguments{
\item{x}{UTC Offsets <e.g., "UTC+09:30" or "UTC-07:00">}
}
\value{
A character vector of standard UTC offsets
}
\description{
Convert numeric values representing UTC offsets (in hours, with
optional fractional parts for minutes) into standardized character
strings of the form `"UTC+HH:MM"` or `"UTC-HH:MM"`. This is useful
for translating numeric time zone offsets into human-readable UTC
offset notation.
}
\examples{
# Convert UTC to numeric values
x <- c (9.5, -7)

x1 <- Num2UTC (x)

print (x1)

}
\seealso{
\code{\link{UTC2Num}}
}
