% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2P.R
\name{R2P}
\alias{R2P}
\title{Convert Traveling Log to Parameters for TAdjust}
\usage{
R2P(Bdf, D, U)
}
\arguments{
\item{Bdf}{A BriefSum object}

\item{D}{The date traveling by plane.}

\item{U}{The UTC offset values <e.g., "UTC+09:30" or "UTC-07:00">}
}
\description{
Convert Traveling Log to Parameters for TAdjust
}
