% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rad2Hr.R
\name{Rad2Hr}
\alias{Rad2Hr}
\title{Rad2Hr}
\usage{
Rad2Hr(x, tau)
}
\arguments{
\item{x}{Numeric vector of phases in radian.}

\item{tau}{Numeric scalar. Period length in the same time units you want returned (for hours use 24). Must be a single numeric value > 0 and <= 24.}
}
\value{
Numeric vector of times in [0, tau). NA values propagate.
}
\description{
Convert `acrophase` in radian to clock time based on tau.
}
\examples{
Rad2Hr(pi/2, tau = 24)

Rad2Hr(c(-pi/2, 0, pi, 3*pi/2), tau = 24)

}
\keyword{acrophase}
\keyword{radian}
