% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SleepWake.R
\name{SleepWake}
\alias{SleepWake}
\title{Classification of Possible Sleep Wake Periods}
\usage{
SleepWake(
  data,
  VAct = NULL,
  VDT = "DateTime",
  method = "CK",
  HyperP = list(CutPoint = NULL, Pad = NA, Scaler = 0.88888, Dist = "ZAG", CutpNW =
    33.33, NDays = 5L, ScreenEpoch = 120L, ArtEpoch = 2L, TraceEpoch = 30L),
  Simple = FALSE,
  Bdf = NULL
)
}
\arguments{
\item{data}{A data.frame of raw actigraphy recording. Both time and activity
count should be included in the \code{data}. See \code{VAct} and \code{VTm}
for further detail.}

\item{VAct}{Optional character. Name of the activity column in \code{data}.
If NULL, defaults to the second column of \code{data}.}

\item{VDT}{Optional character. Name of the \code{POSIXct} datetime column in
\code{data}.If NULL, defaults to "DateTime" of \code{data}.}

\item{method}{Character string specifying detection method. For details on
the available methods and their hyperparameters, see
\itemize{
  \item "CircaCP": A local change-point detection algorithm guided by
  cosinor rhythm, proposed by Chen and Sun (2024).

  \item "CK": A threshold-based classifier proposed by Cole and Kripke
  (1992). The algorithm dichotomies each epoch based on a composite score
  (a sum of weighted activity counts).

  \item "Okley": A threshold-based classifier proposed by Oakley (1997). The
  algorithm is the primary sleep-wake scoring method used in the Actiware
  software by Respironics, Inc.
}}

\item{HyperP}{Optional; a list of hyperparameters used to tune the sleep
wake detection algorithms. The specific hyperparameters depend on the chosen
method. See details below.
\itemize{
  \item "CK":
   \itemize{
     \item CutPoint: Numeric scalar. The threshold for
     classifying sleep vs. wake. Default = 1
     \item Pad: Character scalar. Padding method for edge epochs where the
     full window overhangs. Available methods include:
       \itemize{
       \item NA (default, no padding): edge epochs return NA (i.e.,
       non-classified)
       \item Reflect: pad by reflecting the data at the edges (e.g., y[2],
       y[3], for left edge)
       \item Zero: pad with zeros at the edges
       }
       }

  \item "CircaCP":
   \itemize{
     \item Dist: The assumed distribution (family in regression) of the
     activity counts within a given day. Supported types of distribution
     includes:
       \itemize{
         \item `Exponential` Exponential distribution
         \item `Gamma` Gamma distribution
         \item `Gaussian` Normal distribution
         \item `Poisson` Poisson distribution
         \item `ZAG` Zero-Augmented Gamma distribution (Default)
       }
     \item CutPoint: Numeric scalar between \eqn{[0, 1]}. The
     threshold is used to dichotomize the fitted cosinor curve into active
     vs. inactive periods by CircaCP. Default = 0.2.
     \item CutpNW: Numeric scalar of the maximum percentage of non-wear
     time allowed for a day to be classified. Default = 33.33 (i.e., one
     third of the day).
     \item NDays: Numeric scalar. Minimum numbers of consecutive
     days for classification. Default = 5. Note that for serial processing,
     the function will automatically adjust the threshold down-to two days
     at the minimum.
     \item ScreenEpoch: Numeric scalar. The minimum threshold to classifying
     a segment as probable non-wear time; Default = 120 (minutes per epoch).
     Note, any changes to \code{ScreenEpoch} will be passed onto both
     \code{\link{NonWear}} and \code{\link[CircaCP]{screen_wear}}.
     \item ArtEpoch: Numeric scalar. The maximum interval of spontaneous
     activity (artifact) allowed between two possible non-wear time periods.
     Default = 2 (minutes)
     \item TraceEpoch: Numeric scalar. The minimum time of inactivity for
     the upstream and downstream of the identified spontaneous activity to
     be considered the entire segment as possible non-wear. Default = 30
     (minutes).
     \item Pad: Character scalar. See \code{Pad} in "CK" method above
     for details.
     }

  \item "Oakley":
   \itemize{
     \item Pad: Character scalar. See \code{Pad} in "CK" method above
     for details.
     \item CutPoint: Optional; numeric scalar OR the string "automatic".
     Default = "automatic".
      \itemize{
       \item "automatic": Actiware Auto Wake Threshold computed from data.
       \item Numeric: User defined cutoff value. Note, a numeric value
       should not be wrapped by quotes (i.e., 40 but not \code{"40"}).
       }
     \item Scaler: Optional, a numeric scaler for tuning automatic
     threshold. It has no effect on user-defined cutoff value. Default is
     0.88888 (per Actiware manual).
       }
}}

\item{Simple}{Optional; logical. If `TRUE`, the function returns a vector of
labels from sleep-wake classification results. Note that **NO** additional
computation (e.g., total sleep time) will be performed, either. If `FALSE`
(default), classification results are amended to the input data.}

\item{Bdf}{Optional; a \code{\link{BriefSum}} object containing per-day
metadata for the recording. Note, if jet lag occurred during the recording,
please, update the metadata using \code{\link{TAdjust}} before passing to
this function.}
}
\value{
An amended raw recording data.frame with non-wearer detection results
storing in a new column:
  \code{pSleep_Wake}, which stands for possible sleep or wake label. The
  values are categorical, where
  \itemize{
       \item "S": indicates that the wearer was possibly asleep
       \item "W": indicates that the wearer was possibly awake
       \item NA: indicates non-classified epochs due to algorithm
       constraints (e.g., rejected due to insufficient recording by
       \code{\link[CircaCP]{screen_wear}}).
       }

When \code{Bdf} is provided, the function also returns an updated
\code{BriefSum} object with two additional columns:
\itemize{
  \item Total_Sleep_Time: total duration of possible sleep time per
  day (in minutes).
  \item Wake_Sleep_Ratio: time spent awake-to-sleep-ratio in percentage.
  }
}
\description{
A wrapper function for classifying possible sleep and wake periods based on
activity count. Please, read the \strong{Details} section below for more
information on the implemented algorithms and their hyperparameters.
}
\details{
The following information contains brief description on the sleep-wake
classification algorithms implemented in this function. Currently supported
methods include: "CircaCP", "CK" (i.e., "Cole–Kripke"), and "Okley". For
more details, please refer to the original publications and the references
provided below.

\strong{CircaCP Algorithm}:

This algorithm is a rhythm-based unsupervised classifier. The algorithm
first transforms each daily recording using a nonlinear least square cosinor
model. The fitted model is then used to guide the identification of the
change point (CP) in the activity pattern. Optimal CP is determined using
the modified information criteria and the Calinski–Harabasz criterion.
By default, recordings with a third or more missing values (e.g., non-wear
time or unrecorded time) will be excluded from classification. The algorithm
supports epoch lengths of 15, 30, or 60 seconds. Users may tune the
threshold for classification by changing the default value of
\code{CutPoint} and the distribution of in the \code{HyperP} argument.

\strong{CK (Cole-Kripke) Algorithm}:

This threshold-based supervised classifier detects the possible sleep wake
periods based on sleep index. The index is computed using weighted sum of
rescaled activity counts around a centered window. Assumed weights were
derived by Cole and colleagues using an in-house model selection software.
The algorithm supports epoch lengths of 10, 30, or 60 seconds. Users may
tune the threshold for classification by changing the default value of
\code{CutPoint} in the \code{HyperP} argument.

\strong{Okley Algorithm}:

This threshold-based supervised classifier detects possible sleep wake
periods based on the weighted sum of activity counts around a centered
window. The algorithm supports epoch lengths of 15, 30, 60, or 120 seconds.
Users may tune the threshold for classification by changing the default
value of \code{CutPoint} in the \code{HyperP} argument.
}
\examples{
# Create quick summary of the recording with adjustment for daylight saving.
BdfList <-
  BriefSum (
  data  = FlyEast,
  SR    = 1 / 60,
  Start = "2017-10-24 13:45:00"
  )

data <- BdfList$data

# Apply Cole-Kripke's sleep wake classification
SleepWake (data = data, method = "CK", VAct = "Activity",
HyperP = list (CutPoint = 1))

# Apply Chen's CircaCP sleep wake classification
SleepWake (data = data, method = "CircaCP", VAct = "Activity",
HyperP = list (CutPoint = 0.2))

# Apply Oakley's sleep wake classification
SleepWake (data = data, method = "Oakley", VAct = "Activity",
HyperP = list (CutPoint = "automatic"))


}
\references{
Chen S, Sun X. Validating CircaCP: a generic sleep–wake cycle detection
algorithm for unlabelled actigraphy data. Royal Society Open Science. 2024;
11(5)doi:10.1098/rsos.231468

Cole RJ, Kripke DF, Gruen W, Mullaney DJ, Gillin JC. Automatic Sleep/Wake
Identification From Wrist Activity. Sleep. 1992;15(5):461-469.
doi:10.1093/sleep/15.5.461

Oakley NR. Validation with polysomnography of the Sleepwatch sleep/wake
scoring algorithm used by the Actiwatch activity monitoring system. Mini
Mitter Co. Sleep. 1997;2:1-140.
}
\seealso{
\code{\link[CircaCP]{sleep_detection}}
}
\keyword{sleep}
