% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAdjust.R
\name{TAdjust}
\alias{TAdjust}
\title{Adjust Time Shift based on Travel Log}
\usage{
TAdjust(Bdf, TLog, TZ = NULL, fork = FALSE)
}
\arguments{
\item{Bdf}{`r lifecycle::badge("deprecated")` Use `Meta` instead.}

\item{TLog}{A structured travel log containing date of travel and local time
zone. Use `TravelLog()` to generate template.}

\item{TZ}{The time zone when the recording started. Default is "NULL", which
use the initial geographical location-based time zone indicator specified
in the Bdf. If "local", it will use the local time zone of the computer.
If specified as a time zone string, it will use the specified time zone.}

\item{fork}{Deprecated. Logical, if TRUE, it will use parallel processing
to speed up the computation. Default is FALSE.}

\item{Meta}{A \code{\link{BriefSum}} object containing per-day metadata for
the recording.}
}
\value{
A \code{\link{BriefSum}} object with adjusted data points and time
shift based on travel log.
}
\description{
`TAdjust()` is a function that corrects data points and time-shift based on
travelling log. Note that it is important to ensure that the UTC-offset
value is correct. When in doubt, please use the UTC function or consult IANA
table in the package \code{data(IANA)}. Daylight saving will not be
reassessed as in \code{\link{BriefSum}}.
}
\details{
This function modifies the meta data of the recording generated by
\code{\link{BriefSum}}. Based on the travel log, it will adjust the time
shift based on the UTC offset values provided. If information about local
practice of daylight saving time is provided in the travel log, it will be
preserved by \code{TAdjus} during adjustment. When left unspecified, the
function will make assumption based on the UTC offset via
\code{\link{UTCwDST}}. For latter, daylight saving time is prioritized if
there is any possibility of time change due to daylight saving among time
zones with UTC offsets. Note that while this fallback strategy may not
affect by travels between certain UTCs time zones (e.g., UTC+08:00, +10:00,
-09:30, etc.), it may risk over/under-adjust post-travel recording by an
hour, otherwise. When in doubt, we recommend avoiding comparisons of pre-
and post-travel data.
}
\examples{
\donttest{
# Create quick summary of the recording with adjustment for daylight saving.
## For more information about FlyEast dataset, please refer to the
## documentation of the dataset or the ActiGlobe online tutorials.

BdfList <- BriefSum (
    data = FlyEast,
    SR = 1 / 60,
    Start = "2017-10-19 13:45:00",
    TZ = 'America/New_York'
)

# Extract only the summary report
Bdf <- BdfList$Bdf


# Adjust time shift based on travel log
Bdf.adj <- TAdjust (Bdf  = Bdf,
                    TLog = TLog, # sample travel log
                    TZ   = 'America/New_York')

print (Bdf)
print (Bdf.adj)
}


}
\seealso{
\code{\link{TravelLog}}
}
\keyword{Actigraphy}
\keyword{Adjust}
