% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeFormat.R
\name{TimeFormat}
\alias{TimeFormat}
\title{Detect Possible Time Format}
\usage{
TimeFormat(Time, as.time = FALSE)
}
\arguments{
\item{Time}{A character string of time.}

\item{as.time}{A binary operator indicating whether to return a converted time based on the detection or the time format. (default: FALSE, which returns time format)}
}
\value{
character
}
\description{
This function will automatically detect possible time format. Users can choose to either reformat the date or report the datetime format detected.
}
\examples{
# Example 1: When all dates have the same format
## Create and store a date in a variable called Time
Time = c("2017/05/02 23:00:01", "1970/01/02 05:10:33", "2000/02/28 07:00:00")

## Ask TimeFormat to tell us the format of the time.
print(TimeFormat(Time))

# Example 2: When multiple formats co-exist in a variable
## Create and store dates and time in a variable called Time
Time = c("2017/05/02 23:00:01", "2000/02/28 07:00", "1970/01/02", "2022/11/28 08:35 PM")

## Ask TimeFormat to tell us the format of the Time.
### DO This!
#### Option 1.
print(lapply(Time,TimeFormat))

#### Option 2. To match the warning message to the items
for(x in Time){

print(TimeFormat(x))

}


### DO NOT!
print(TimeFormat(Time))
### Note, this process will fail because there are multiple formats




}
