% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TravelLog.R
\name{TravelLog}
\alias{TravelLog}
\title{Generate the Travel Log Template Needed for Time Zone Correction}
\usage{
TravelLog(Type = "UTC", Write = FALSE, Dir = NULL)
}
\arguments{
\item{Type}{A character string to indicate the type of template to create.
Options are
\itemize{
 \item "UTC": the default template that requires users to input the UTC
 offset (e.g., "UTC +09:30" or "+9.5") of the local or travel destination
 time zone.
 \item "CC": a template that requires users to input both city and country
 names of the location. This template is not suited for analysis involving
 sensitive information. In rare cases that multiple cities share the same in
 one country, users should input the correct corresponding state or province
 names.
}}

\item{Write}{A binary code to indicate whether to write a .csv file
containing the template needed for the travel log. (default = FLASE)
When set to FALSE, the template will be returned as an object.
When set to TRUE, user must provide the target directory where they wish to
store the template.}

\item{Dir}{The directory where the travel log template to be exported <e.g.
"C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
}
\value{
a travel-log template as a data.frame or written as a CSV file
}
\description{
A helper function to generate a travel log template needed for time zone
correction. This function creates a data frame with the necessary columns
for users to fill in their travel log information. This is an essential
step for analysis involving time zone shifts. We recommend users to consult
the package vignette to learn how to use the travel log template.
}
\details{
The travel log template can be generated in two formats: "UTC" and "CC".
The "UTC" (coordinated universal time) template is general recommended as it
aims to protect personal privacy of participants without compromising
reproducibility. Note that the "CC" (city-country) template should be
avoided when analyzing data containing sensitive information.

For UTC offset, users may consult the IANA database \code{View(IANA)},
already installed with the package. Alternatively, users may also use
\code{\link{TZ2UTC}} or \code{\link{Geo2TZ}} to obtain the UTC offset.

Spelling check is recommended when using the "CC" template, as mistakes can
halt the process of time shift adjustment. It is recommended to keep the
documentation in English since accents may not be recognized by the default
encoding system depending on the operating system.
}
\examples{

Tlg <- TravelLog (Write = FALSE)

print (Tlg)

}
\keyword{Log}
\keyword{Shift}
\keyword{Template}
\keyword{Timezone}
\keyword{Travel}
