% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TZ2UTC.R
\name{TZ2UTC}
\alias{TZ2UTC}
\title{Compute UTC offset based on Time Zone and Date}
\usage{
TZ2UTC(Date = NULL, TZ = "local", DT = NULL)
}
\arguments{
\item{Date}{A POSIXct date used as the reference point for offset
comparison. Defaults to January 1, 2021 UTC if `NULL`.}

\item{TZ}{The time zone when the recording started. Default = "local",
which means user's local time zone}

\item{DT}{`r lifecycle::badge("deprecated")` Use `Date` instead.}
}
\value{
A character string indicating the UTC offset, e.g., "UTC-05:00"
}
\description{
Compute UTC offset based on Time Zone and Date
}
\examples{
x <- as.Date(c("2017-10-24", "2017-11-20"))
TZ2UTC(Date = x, TZ = "America/New_York") ## Expect two different UTCs


# A vector of dates
x <- seq.Date(from = as.Date("2017-03-15"),
              to   = as.Date("2017-11-27"),
              by   = "day")


# If the user resides in a country that follows daylight saving time
TZ2UTC(Date = x, TZ = "local")

\dontrun{
# If not
 TZ2UTC(Date = x, TZ = "America/New_York")
}

}
\keyword{TZ}
\keyword{UTC}
