% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TZwDST.R
\name{TZwDST}
\alias{TZwDST}
\title{Find DST Start and End Date/Time for IANA Time Zones}
\usage{
TZwDST(TZ, Year = NULL, Simple = TRUE)
}
\arguments{
\item{TZ}{A character vector of time zone identifiers.}

\item{Year}{Integer. Defaults to current system year if NULL.}

\item{Simple}{Logical. If TRUE (default), returns a logical vector
indicating whether DST occurs in each time zone for the specified year.
If FALSE, returns a detailed data frame with DST transition information.}
}
\value{
Two possible outputs based on the `Simple` parameter:

A logical vector of the same length as `TZ`.
  \itemize{
    \item `TRUE` Daylight saving time occurs during the year.
    \item `FALSE` No daylight saving time during the year.
}

 A data.frame with one row per time zone and columns:
\itemize{
  \item Information on Daylight Saving Time Transition
  \itemize{
   \item TZ_NAME: Time zone name.
   \item UTC_DST: UTC offset during DST (e.g., "UTC -04:00").
   \item DST_Start_Date: Date when DST starts (NA if no DST).
   \item DST_Start_Time: Time when DST starts (NA if no DST).
   \item UTC_Standard: UTC offset during standard time (e.g., "UTC -05:00").
   \item DST_End_Date: Date when DST ends (e.g, "2021-11-07", NA if no DST).
   \item DST_End_Time: Time when DST ends (e.g., "02:00:00", NA if no DST).
   }
  \item Time Zone Shift (Special Cases)
  \itemize{
   \item Original_UTC: UTC offset before day jump* or time zone shift or
   realignment (e.g., "Australia/LHI"  in 1981).
   \item Shift_Start_Date: Date when the shift starts.
   \item Shift_Start_Time: Time when the shift starts.
   \item New_UTC: UTC offset after the shift.
   \item New_Date: Date after the shift.
   \item New_Time: Time after the shift.
   }
}
}
\description{
`TZwDST` identifies the start and end date/time of daylight saving
time (DST) for a given set of IANA time zones and a specified year.
It returns a data frame summarizing the DST transitions and UTC offsets
for each time zone.
}
\examples{
# Find DST transitions for a few time zones in 2021
TZwDST(TZ = c("America/New_York", "Europe/London
", "Asia/Tokyo"), Year = 2021)

TZwDST(TZ = "Africa/Cairo", Year = 2010)

\donttest{
# Find DST transitions for all time zones in the current year
TZwDST(TZ = OlsonNames()) # This will take a while
}

}
\seealso{
\code{\link{OlsonNames}}
}
\keyword{dst}
\keyword{tz}
\keyword{year}
