% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTC2Num.R
\name{UTC2Num}
\alias{UTC2Num}
\title{Convert Coordinated Universal Time (UTC) Offset to Numeric Values}
\usage{
UTC2Num(x)
}
\arguments{
\item{x}{Character string of UTC offsets, e.g., "UTC+09:30" or "UTC-07:00".}
}
\value{
Numeric values of UTC offsets
}
\description{
Convert Coordinated Universal Time (UTC) Offset to Numeric Values
}
\examples{
# Convert UTC to numeric values
x <- c ("UTC+09:30", "UTC-07:00")

x1 <- UTC2Num (x)

print (x)


}
\seealso{
\code{\link{Num2UTC}}
}
\keyword{UTC}
\keyword{conversion}
\keyword{offset}
