% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.act.R
\name{write.act}
\alias{write.act}
\title{Export Daily Activity Score and Recording Summary}
\usage{
write.act(
  Dir,
  df,
  Bdf,
  ID,
  TUnit = "hour",
  VAct = NULL,
  VTm = NULL,
  Incomplete = FALSE
)
}
\arguments{
\item{Dir}{The directory where the recordings to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{df}{A data.frame containing the actigraphy records}

\item{Bdf}{A BriefSum object (default NULL, i.e., not included). This is essential when time adjustment is needed.}

\item{ID}{The subject's ID which would be used to create a folder.}

\item{TUnit}{The unit of time to be displayed in the daily actigraphy recording exported. These include: "day" "hour" "minute" "second"  (default: "hour")}

\item{VAct}{The name of the variable containing activity score  (default assumed the second variable)}

\item{VTm}{The name of the variable containing time stamp (default assumed the first variable)}

\item{Incomplete}{A binary operator to decide whether to keep any less than 24hr-long recordings. (default: FALSE)}
}
\description{
A function that exports two types of files locally. 1. A summary of the entire recording (.csv). 2. Daily activity recordings in individual files (.txt). The default deliminator is set to be space, with `.` as the decimal indicator. As such, it should be noted that some transformation may be needed when opening in EXCEL on a computer with non-English EU settings.
}
\keyword{Actigraphy}
\keyword{Export}
