% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.cosinor.R
\name{write.cosinor}
\alias{write.cosinor}
\title{Generate Report for Daily Cosinor and Generate Summary Plot}
\usage{
write.cosinor(Dir, ID, DailyAct, Bdf, VAct = NULL, VTm = NULL)
}
\arguments{
\item{Dir}{The directory where the recordings to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{ID}{The subject's ID which would be used to create a folder.}

\item{DailyAct}{A data.frame list generated by `Act2Daily()`}

\item{Bdf}{The summary report created by BriefSum.}

\item{VAct}{The name of the variable containing activity score  (default assumed the second variable)}

\item{VTm}{The name of the variable containing the time coordiante (default assumed the first variable)}
}
\description{
`write.cosinor()` is a multi-wrapper function which generate cosinor plot using CosinorM and data trend using `geom_smooth`. All daily results could be found in the exported report.
}
\examples{
\dontrun{

# Import data
data(FlyEast)


# Create quick summary of the recording with adjustment for daylight saving.
BdfList <-
BriefSum(df = FlyEast ,
         SR = 1/60,
         Start = "2017-10-24 13:45:00")


 # Let's extract the quick summary of the recording
Bdf <- BdfList$Bdf

### Note that since the original data was affected by travel-induced time
### shift, the recordings would not be properly segmented from 2017-11-02.

## To avoid time shift due to travelling, we will keep only the first 8 days.
Bdf <- Bdf[1:8,]


## Segment Data by Day
dfList =
 Act2Daily(df = FlyEast,
           Bdf = Bdf,
           VAct = "Activity",
           VTm = "X2",
           Incomplete = TRUE,
           Travel = TRUE)


## Generate graphic summary of daily recordings in pdf with an addition
## `Summary` file just like write.act
 write.cosinor(Dir = getwd(), ## Export to the current working directory
               ID = "JD",
               df = dfList$Daily_df,
               Bdf = Bdf,
               VAct = "Activity",
               VTm = "Time")

## To see what was exported copy the path generated by the code below and
## paste it as an url in a web browser or the file explorer.
 getwd()


}
}
\seealso{
\code{\link{Act2Daily}}
}
\keyword{Actigraphy}
\keyword{Cosinor}
\keyword{Export}
\keyword{Trend}
