% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Act2Daily.R
\name{Act2Daily}
\alias{Act2Daily}
\title{Split and Annotate Actigraphy Data into Daily Records}
\usage{
Act2Daily(
  df,
  Bdf,
  TUnit = "hour",
  VAct = NULL,
  VTm = NULL,
  Incomplete = FALSE,
  Travel = TRUE
)
}
\arguments{
\item{df}{A data.frame of raw actigraphy recording. Both time and activity count
should be included in the `df`. See `VAct` and `VTm` for further detail.}

\item{Bdf}{A BriefSum object. If jet lag or daylight saving occurred during the recording,
please, use the updated version from \code{TAdjust}.}

\item{TUnit}{Character; time‐unit for the x‐axis of each day's timeline.
Must be one of `"day"`, `"hour"`, `"minute"` or `"second"`.  Default is `"hour"`.}

\item{VAct}{Optional character.  Name of the activity column in `df`. If NULL,
defaults to the second column of `df`.}

\item{VTm}{Optional character.  Name of the time index column in `df`. If NULL,
defaults to the first column of `df`.}

\item{Incomplete}{Logical; if TRUE, days flagged `"Incomplete Recording"` (i.e. <24 h)
are retained in the data list with recordings segmented by day. Default = FALSE (these days are removed).}

\item{Travel}{Logical; if TRUE, days flagged `"Travel"` are retained although
some data points from an earlier adjacent calendar day may be duplicated (a
warning is issued). If FALSE, travel days and the day before/after are excluded.
Default = TRUE.}
}
\value{
A data list that includes the adjusted recordings both in long format and as a nested list by day.
}
\description{
`Act2Daily` is used to split the actigraphy recording by the recording dates stored in Bdf (a BriefSum object).
}
\details{
Transforms a continuous actigraphy time‐series into a day‐by‐day list
 of aligned epoch and produces a fully annotated data frame. Based on the
 summary table generated by either [BriefSum()] or [TAdjust()], it re-organize and
 segment the raw recording data by each day. The function is designed to
 process incomplete recordings while accounting for travel and daylight‐saving
 adjustments. All labels contained in the summary table (i.e., Bdf) will be
 assigned to the corresponding data points. This workflow ensures that each day's recording is bounded by its
 calendar date while accounts for travel‐induced overlaps, and clearly documents any incomplete or `unallocated` points.
}
\note{
When you cross into a time zone with a larger UTC offset `(e.g., Montreal UTC–5 → Taipei UTC+8)`, it will generate a small overlap of epochs between the day
before travel and the day labeled `“Travel Day”` in the `"Daily_df"`. Conversely,
when you move into a time zone with a smaller UTC offset, a brief gap appears
on `“Travel Day”`, with labeling - `"Unallocated"`, which represents "time lose".
}
\examples{
# Option 1. Create a list that includes the adjusted recordings both in long
# format and as a nested list by day.

\dontrun{
dfList =
 Act2Daily(df = FlyEast,
           Bdf = Bdf.adj,
           VAct = "Activity",
           VTm = "X2",
           Incomplete = TRUE,
           Travel = TRUE)

}

# Option 2, separate storage with the help of "\%<-\%" operate from "zeallot"
\dontrun{
library(zeallot)
c(Dailydf, Newdf) \%<-\%
 Act2Daily(df = df,
           Bdf = Bdf.adj,
           VAct = "Activity",
           VTm = "X2",
           Incomplete = TRUE,
           Travel = TRUE)
}
}
\keyword{Actigraphy}
\keyword{Daily}
\keyword{Segment}
