% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BriefSum.R
\name{BriefSum}
\alias{BriefSum}
\title{Summarize the Actigraphy Recording by Day}
\usage{
BriefSum(df, SR, Start, TZ = "local")
}
\arguments{
\item{df}{A data.frame containing the actigraphy records}

\item{SR}{The sampling rate of the actigraphy (unit at Hz). Note that Hz should be equal to or less than 1.}

\item{Start}{The starting date and time of the recording in formats like this "2021-03-05 18:31:03"}

\item{TZ}{The time zone when the recording started. (default = "local", which means user's local time zone)}
}
\value{
A data.frame where each row holds all metadata for one recording day: the calendar date, time‐zone code, epoch length (seconds), UTC offset, daylight‐saving flag, cumulative start/end seconds from midnight, recording start/end times (HH:MM:SS), any warning labels (e.g. “Travel”, “Incomplete Recording”), an exclusion flag, and the expected number of epochs for a full day
}
\description{
Summarize the Actigraphy Recording by Day
}
\examples{
# Export the data in a list
\dontrun{
data(FlyEast)

Bdf = BriefSum(df = FlyEast,
               SR = 1/60,
               Start = "2017-10-19 13:45:00")
print(Bdf)

## install library "zeallot"
## library(zeallot)
## c(Bdf, df) \%<-\%
## BriefSum(df = FlyEast,
##          SR = 1/60,
##          Start = "2017-10-24 13:45:00")

}

}
\keyword{Actigraphy}
\keyword{Summary}
