% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateFormat.R
\name{DateFormat}
\alias{DateFormat}
\title{Detect Possible Datetime Format}
\usage{
DateFormat(DateTime, as.date = TRUE, Delim = NULL)
}
\arguments{
\item{DateTime}{A character string of date.}

\item{as.date}{A binary operator indicating whether to return a converted date based on the detection or the date time format. (default: TRUE)}

\item{Delim}{If the deliminator used to separate month and date is not of standard  (i.e., / or . or -), please specify it here. (default: NULL)}
}
\value{
character
}
\description{
This function will automatically detect possible date format. Users can choose to either reformat the date or report the datetime format detected.
}
\examples{
# Example 1: When all dates have the same format
## Create and store a date in a variable called DateTime
DateTime = c("2017/05/02", "2000/02/28", "1970/01/02")

## Ask DateFormat to tell us the format of the date.
print(DateFormat(DateTime))

# Example 2: When multiple formats co-exist in a variable
## Create and store a date in a variable called DateTime
DateTime = c("2017/05/02", "2000.Feb.28", "1970-11-02", "January 01, 2025", "December 12, 1980")

## Ask DateFormat to tell us the format of the date.
### DO This!
#### Option 1.
print(lapply(DateTime,DateFormat))

#### Option 2. To match the warning message to the items
for(x in DateTime){

print(DateFormat(x))

}


### DO NOT!
print(DateFormat(DateTime))
### Note, this process will fail because there are multiple formats




}
