% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggActiGlobe.R
\name{ggActiGlobe}
\alias{ggActiGlobe}
\title{Plot an Overview of an `ActiGlobe` Activity Time Series}
\usage{
ggActiGlobe(df, Bdf, VAct, VDT = "DateTime", VD = "Date", ...)
}
\arguments{
\item{df}{A data.frame of annotated actigraphy epochs.  Must include:
- An activity column named by `VAct`.
- A datetime column named by `VDT`.
- Optionally, a `Note` column to flag affected epochs.}

\item{Bdf}{A BriefSum object containing per‐day metadata for the recording.
If you have applied jet‐lag or daylight‐saving adjustments, pass the output
from the ActiGlobe function `TAdjust`.}

\item{VAct}{Character. Name of the activity column in `df` (numeric counts).}

\item{VDT}{Character. Name of the POSIXct datetime column in `df`.}

\item{...}{graphical parameters to plot}
}
\value{
A `ggplot` object showing:
  - Activity counts vs. time.
  - Dashed vertical lines at each midnight.
  - Points `colored` by whether they were flagged in `Note`.
}
\description{
Creates a time‐series `scatterplot` of activity counts from an `ActiGlobe` data.frame,marking each midnight boundary with a vertical dashed line and `coloring` points that were flagged (e.g. travel overlaps or unallocated epochs).
}
\examples{
\dontrun{
library(ActiGlobe)

# Overview the Uncorrected Longitudinal Recording
data(FlyEast)

BdfList =
BriefSum(df = FlyEast ,
         SR = 1/60,
         Start = "2017-10-24 13:45:00")

p <- ggActiGlobe(df   = BdfList$df,
                 Bdf  = BdfList$Bdf,
                 VAct = "Activity",
                 VDT  = "DateTime")

print(p)



# Overview the Corrected Longitudinal Recording
data(TLog)

BdfList$Bdf.adj = TAdjust(BdfList$Bdf, TLog)
p2 <-  ggActiGlobe(df   = BdfList$df,
                 Bdf  = BdfList$Bdf,
                 VAct = "Activity",
                 VDT  = "DateTime")
print(p2)

# Pro-tip: [`cowplot`] can help stack the time series graphs in one single plot
}

}
\keyword{actigraphy}
\keyword{visualization}
