% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAdjust.R
\name{TAdjust}
\alias{TAdjust}
\title{Adjust Time Shift based on Travel Log}
\usage{
TAdjust(Bdf, TLog)
}
\arguments{
\item{Bdf}{A BriefSum object in data.frame}

\item{TLog}{A structured travel log containing date of travel and local time zone. Use `TravelLog()` to generate template.}
}
\description{
`TAdjust()` is a function that corrects data points and time-shift based on traveling log.
}
\seealso{
[TravelLog()]
}
\keyword{Actigraphy}
\keyword{Summary,}
