% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAdjust.R
\name{TAdjust}
\alias{TAdjust}
\title{Adjust Time Shift based on Travel Log}
\usage{
TAdjust(Bdf, TLog, TZ = NULL)
}
\arguments{
\item{Bdf}{A BriefSum object in data.frame}

\item{TLog}{A structured travel log containing date of travel and local time zone. Use `TravelLog()` to generate template.}

\item{TZ}{The time zone when the recording started. (default = "NULL", which will disregard the use of the initial geographical location-based time zone indicator)}
}
\description{
`TAdjust()` is a function that corrects data points and time-shift based on travelling log. Note that it is important to ensure that the UTC-offset value is correct. When in doubt, please use the UTC function or consult IANA table in the package. Daylight saving will not be reassessed as in BriefSum function.
}
\seealso{
[TravelLog()]
}
\keyword{Actigraphy}
\keyword{Adjust}
