% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAdjust.R
\name{TAdjust}
\alias{TAdjust}
\title{Adjust Time Shift based on Travel Log}
\usage{
TAdjust(Bdf, TLog, TZ = NULL, fork = FALSE)
}
\arguments{
\item{Bdf}{A \code{\link{BriefSum}} object containing per-day metadata for the recording.}

\item{TLog}{A structured travel log containing date of travel and local time zone. Use `TravelLog()` to generate template.}

\item{TZ}{The time zone when the recording started. (default = "NULL", which will disregard the use of the initial geographical location-based time zone indicator)}

\item{fork}{Logical, if TRUE, it will use parallel processing to speed up the computation. Default is FALSE.}
}
\value{
A \code{\link{BriefSum}} object with adjusted data points and time shift based on travel log.
}
\description{
`TAdjust()` is a function that corrects data points and time-shift based on travelling log. Note that it is important to ensure that the UTC-offset value is correct. When in doubt, please use the UTC function or consult IANA table in the package. Daylight saving will not be reassessed as in BriefSum function.
}
\examples{
\dontrun{

# Import sample data
data (FlyEast)

# Create quick summary of the recording with adjustment for daylight saving.
BdfList <- BriefSum (
    df = FlyEast,
    SR = 1 / 60,
    Start = "2017-10-19 13:45:00"
)

# Extract only the summary report
Bdf <- BdfList$Bdf

# Import sample travel Log
data (TLog)

# Adjust time shift based on travel log
Bdf.adj <- TAdjust (Bdf, TLog)

# Display the summary
View (Bdf)
View (Bdf.adj) ### Focus on the dates after 2017-11-01
}

}
\seealso{
\code{\link{TravelLog}}
}
\keyword{Actigraphy}
\keyword{Adjust}
