% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TravelLog.R
\name{TravelLog}
\alias{TravelLog}
\title{Generate the Travel Log Template Needed for Time Zone Correction}
\usage{
TravelLog(Write = FALSE, Dir = NULL)
}
\arguments{
\item{Write}{A binary code to indicate whether to write a .csv file containing the template needed for the travel log. (default = FLASE)
When set to FALSE, the template will be returned as an object.
When set to TRUE, user must provide the target directory where they wish to store the template.}

\item{Dir}{The directory where the travel log template to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
}
\value{
data.list
}
\description{
Generate the Travel Log Template Needed for Time Zone Correction
}
\keyword{Log}
\keyword{Shift}
\keyword{Template}
\keyword{Timezone}
\keyword{Travel}
