% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TZ2UTC.R
\name{TZ2UTC}
\alias{TZ2UTC}
\title{Compute UTC offset based on Time Zone and Date}
\usage{
TZ2UTC(DT, TZ = "local")
}
\arguments{
\item{DT}{The date in the format as "2021-03-05"}

\item{TZ}{The time zone when the recording started. (default = "local", which means user's local time zone)}
}
\description{
Compute UTC offset based on Time Zone and Date
}
\examples{

x <- as.Date(c("2017-10-24","2017-11-20"))
TZ2UTC(DT = x, TZ = "America/Toronto") ## Expect two different UTCs

\dontrun{
# A vector of dates
x =
seq.Date(from = as.Date("2017-10-24"),
           to = as.Date("2017-11-27"),
           by = "day")

# If the user resides in a country that follows daylight saving time
TZ2UTC(DT = x, TZ = "local")

# If not
# TZ2UTC(DT = x, TZ = "America/Toronto")
}

}
\keyword{Daylight}
\keyword{Saving}
\keyword{TZ}
\keyword{Timezone}
\keyword{UTC}
