% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTC2Num.R
\name{UTC2Num}
\alias{UTC2Num}
\title{Convert Standard UTC Offset to Numbers}
\usage{
UTC2Num(x)
}
\arguments{
\item{x}{UTC Offsets <e.g., "UTC+09:30" or "UTC-07:00">}
}
\description{
Convert Standard UTC Offset to Numbers
}
\examples{
# Convert UTC to numeric values
x = c("UTC+09:30","UTC-07:00")

x1 = UTC2Num(x)

print(x)

}
