% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.act.R
\name{write.act}
\alias{write.act}
\title{Export Daily Activity Score and Recording Summary}
\usage{
write.act(
  Dir,
  ID,
  df,
  Bdf,
  TUnit = "hour",
  VAct = NULL,
  VTm = NULL,
  Incomplete = FALSE,
  Travel = TRUE,
  Simple = FALSE
)
}
\arguments{
\item{Dir}{The directory where the recordings to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{ID}{The subject's ID which would be used to create a folder.}

\item{df}{A data.frame of raw actigraphy recording. Both time and activity count
should be included in the `df`. See `VAct` and `VTm` for further detail.}

\item{Bdf}{A \code{\link{BriefSum}} object. If jet lag or daylight saving occurred during the recording,
please, use the updated version from \code{\link{TAdjust}}.}

\item{TUnit}{Character; time--unit for the x--axis of each day's timeline.
Must be one of `day`, `hour`, `minute` or `second`.  Default is `hour`.}

\item{VAct}{Optional character.  Name of the activity column in `df`. If NULL,
defaults to the second column of `df`.}

\item{VTm}{Optional character.  Name of the time index column in `df`. If NULL,
defaults to the first column of `df`.}

\item{Incomplete}{Logical; if TRUE, days flagged `Incomplete Recording` (i.e. <24 h)
are retained in the data list with recordings segmented by day. Default = FALSE (these days are removed).}

\item{Travel}{Logical; if TRUE, days flagged `Travel` are retained although
some data points from an earlier adjacent calendar day may be duplicated (a
warning is issued). If FALSE, travel days and the day before/after are excluded.
Default = TRUE.}

\item{Simple}{Logical; if TRUE, only columns stored in the original recordings will be exported.
If FALSE, all information stored in `df` will be generated. Default = FALSE}
}
\value{
Text files ".txt" containing segmented recordings of each day and `Bdf` from \code{\link{BriefSum}}

\itemize{
  \item Text File: Recordings segmented by the recording date
  \item Summary: Bdf exported as ".csv" file.
  }
}
\description{
A function that exports two types of files locally. 1. A summary of the entire recording (.csv). 2. Daily activity recordings in individual files (.txt). The default deliminator is set to be space, with `.` as the decimal indicator. As such, it should be noted that some transformation may be needed when opening in EXCEL on a computer with non-English EU settings.
}
\examples{

\dontrun{

# Import data
data(FlyEast)


# Create quick summary of the recording with adjustment for daylight saving.
BdfList <-
BriefSum(df = FlyEast ,
         SR = 1/60,
         Start = "2017-10-24 13:45:00")

 # Let's extract the quick summary of the recording
Bdf <- BdfList$Bdf

### Note that since the original data was affected by travel-induced time
### shift, the recordings would not be properly segmented from 2017-11-02.

## To avoid time shift due to travelling, we will keep only the first 8 days.
Bdf <- Bdf[1:8,]




# Export daily recordings and the summary report

 write.act(Dir = getwd(), ## Export to the current working directory
           ID = "JD",
           df = FlyEast,
           Bdf = Bdf,
           VAct = "Activity",
           VTm = "X2")

## To see what was exported copy the path generated by the code below and
## paste it as an url in a web browser or the file explorer.
 getwd()

}


}
\keyword{Actigraphy}
\keyword{Export}
