% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.cosinor.R
\name{write.cosinor}
\alias{write.cosinor}
\title{Generate Report for Daily Cosinor and Generate Summary Plot}
\usage{
write.cosinor(Dir, DailyAct, Bdf, ID, VAct = NULL, VTm = NULL)
}
\arguments{
\item{Dir}{The directory where the recordings to be exported <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{DailyAct}{A data.frame list generated by `Act2Daily()`}

\item{Bdf}{The summary report created by BriefSum.}

\item{ID}{The subject's ID which would be used to create a folder.}

\item{VAct}{The name of the variable containing activity score  (default assumed the second variable)}

\item{VTm}{The name of the variable containing the time coordiante (default assumed the first variable)}
}
\description{
`write.cosinor()` is a multi-wrapper function which generate cosinor plot using CosinorM and data trend using `geom_smooth`. All daily results could be found in the exported report.
}
\seealso{
[Act2Daily()]
}
\keyword{Actigraphy}
\keyword{Cosinor}
\keyword{Export}
\keyword{Trend}
